petstore_v2_app/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-eventgrid
azure-core
cloudevents
azure-storage-blob
pymssql